Remove the scheduler, was causing more trouble compiling the damn thing than it was worth

+++ src/cpu8086/mod.rs
@@ -96,7 +96,7 @@ impl Cpu8086 {
         self.mem_read_word(ctx, self.regs.readseg16(SegReg::SS), stack_pointer)
     }
 
-    pub fn tick<T: Cpu8086Context>(&mut self, ctx: &mut T) -> u128 {
+    pub fn tick<T: Cpu8086Context>(&mut self, ctx: &mut T) -> usize {
         self.opcode = self.mem_read_byte(ctx, self.regs.readseg16(SegReg::CS), self.regs.ip);
         println!(
             "Opcode {:#02x} CS {:#04x} IP {:#04x}\nGPRs {:x?} FLAGS {:#04x}",
